+++ src/js/core/switcher.js
@@ -148,27 +148,21 @@ export default {
 
         show(item) {
             const prev = this.index();
-            const next = getIndex(
-                this.children[getIndex(item, this.toggles, prev)],
-                children(this.$el)
-            );
-
-            if (prev === next) {
-                return;
-            }
-
+            const next = getIndex(item, this.toggles, prev);
+            const active = getIndex(this.children[next], children(this.$el));
             children(this.$el).forEach((child, i) => {
-                toggleClass(child, this.cls, next === i);
-                attr(this.toggles[i], 'aria-expanded', next === i);
+                toggleClass(child, this.cls, active === i);
+                attr(this.toggles[i], 'aria-expanded', active === i);
             });
 
+            const animate = prev >= 0 && prev !== next;
             this.connects.forEach(async ({ children }) => {
                 await this.toggleElement(
                     toNodes(children).filter((child) => hasClass(child, this.cls)),
                     false,
-                    prev >= 0
+                    animate
                 );
-                await this.toggleElement(children[next], true, prev >= 0);
+                await this.toggleElement(children[active], true, animate);
             });
         },
     },